+++ Diagnostics/Assets/Scripts/Protocols/Protocol.cs
@@ -9,12 +9,14 @@ namespace Protocols
         public string Title { get; set; }
         public Appearance Appearance { get; set; }
         public string Introduction { get; set; }
+        public bool FullAuto {  get; set; }
 
         public List<ProtocolEntry> Tests { get; set; }
 
         public Protocol()
         {
             Appearance = new Appearance();
+            FullAuto = false;
         }
 
     }